Task1 modification by tl

+++ GITProj1/src/main/java/com/nt/service/Arthmetic.java
@@ -2,7 +2,7 @@ package com.nt.service;
 
 public class Arthmetic
 {
-	public int sun(int x, int y)
+	public double sun(double x, double y)
 	{
 		return x+y;
 	}
